+++ src/index.ts
@@ -1,5 +1,5 @@
-export type {TrackedInstance} from './tracked-instance'
-export type {Collection, CollectionItem} from './collection'
+export type {TrackedInstance, TrackedInstanceOptions} from './tracked-instance'
+export type {Collection, CollectionItem, CollectionOptions} from './collection'
 
 export {useTrackedInstance} from './tracked-instance'
 export {useCollection} from './collection'